requirements.txt
Flask==1.1.1
flask_mongoengine==0.9.3
py-bcrypt==0.4
Flask-Login==0.5.0
google-cloud==0.29.0
Flask-DotEnv==0.1.2
python-dotenv==0.12.0